ci: Solve circular dependencies on utils
