+++ apps/main-landing/src/app/creators/utils/index.ts
@@ -32,136 +32,6 @@ export const browserBasedSource = ({
   return svgSrc;
 };
 
-export const getPublicCreatorProfile = async (
-  name?: string | null,
-  address?: Hex | null,
-): Promise<CreatorProfileResponse | undefined> => {
-  if (!name && !address) {
-    console.error('No name or address to get creator');
-    return;
-  }
-
-  const lookPath = address ? `address/${address}` : `${name}`;
-
-  const response = await fetch(
-    `${CREATOR_API_URL}/creator-profile/${lookPath}`,
-  );
-  if (!response.ok) {
-    return;
-  }
-  const data = (await response.json()) as CreatorProfileResponse;
-  return data;
-};
-
-export const getPublicCreatorProfileBySlug = async (
-  slug: string | null,
-): Promise<CreatorProfileResponse | undefined> => {
-  if (!slug) {
-    console.error('No slug to get creator');
-    return;
-  }
-
-  const response = await fetch(
-    `${CREATOR_API_URL}/creator-profile/donation-overlay/${slug}`,
-  );
-  if (!response.ok) {
-    return;
-  }
-  const data = (await response.json()) as CreatorProfileResponse;
-  return data;
-};
-
-export const saveCreatorProfile = async (
-  address: Hex,
-  name?: string | null,
-  displayName?: string | null,
-  profilePictureUrl?: string | null,
-  email?: string | null,
-  privyId?: string | null,
-  authToken?: string,
-  isDonor?: boolean,
-): Promise<void> => {
-  if (!address || !name || !privyId) {
-    throw new Error('No wallet address, name or privyId to create creator');
-  }
-
-  if (!authToken) {
-    throw new Error('No auth token provided');
-  }
-
-  const response = await fetch(`${CREATOR_API_URL}/creator-profile`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      'Authorization': `Bearer ${authToken}`,
-    },
-    body: JSON.stringify({
-      address: address,
-      primaryAddress: address,
-      displayName,
-      profilePictureUrl,
-      name,
-      email,
-      privyId,
-      isDonor,
-    }),
-  });
-
-  if (!response.ok) {
-    throw new Error('Failed to register creator');
-  }
-};
-
-export const editCreatorProfile = async (
-  name: string,
-  profile: Partial<CreatorProfileResponse>,
-  authToken?: string,
-): Promise<boolean> => {
-  try {
-    if (!name) {
-      console.error('No name provided to edit creator profile');
-      return false;
-    }
-
-    if (!authToken) {
-      console.error('No auth token provided');
-      return false;
-    }
-
-    const response = await fetch(`${CREATOR_API_URL}/creator-profile/${name}`, {
-      method: 'PATCH',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${authToken}`,
-      },
-      body: JSON.stringify({
-        ...profile,
-      }),
-    });
-
-    if (!response.ok) {
-      return false;
-    }
-    return true;
-  } catch (error) {
-    console.error('Error updating creator profile:', error);
-    return false;
-  }
-};
-
-export const deleteCreatorAccount = async (authToken: string) => {
-  const response = await fetch(`${CREATOR_API_URL}/creator-profile/me`, {
-    method: 'DELETE',
-    headers: {
-      'Content-Type': 'application/json',
-      'Authorization': `Bearer ${authToken}`,
-    },
-  });
-  if (!response.ok) {
-    throw new Error('Deletion failed');
-  }
-};
-
 export const getChainShortNamesFromIds = (chainsIds: number[]) => {
   return (
     chainsIds
@@ -259,5 +129,14 @@ export async function getNftMetadata(
   }
 }
 
-export { setCreatorIfSessionPresent } from './session';
-export { useStartEarningNavigation, getCreatorProfile } from './navigation';
+export {
+  removeDonorStatus,
+  setCreatorIfSessionPresent,
+  getCreatorProfile,
+  getPublicCreatorProfile,
+  getPublicCreatorProfileBySlug,
+  saveCreatorProfile,
+  editCreatorProfile,
+  deleteCreatorAccount,
+} from './session';
+export { useStartEarningNavigation } from './navigation';
